Don't throw on lazy item updating unexpanded globs
An Item `Update` operation may apply to items that are defined via
wildcard. In normal operation, the wildcard will have been expanded by
the time Update applies, and it's fine for Update to assume that
everything is a valid file path. But in
`MSBuildSkipEagerWildCardEvaluationRegexes` mode, wildcards may not be
expanded.

Fixes #9405 by using a more appropriate method to normalize
probably-but-not-necessarily paths for comparisons.